+++ sw-mcu/canfd-to-usb-nucleo/Core/Src/utlFDCAN_Task.c
@@ -6,11 +6,6 @@
  */
 #include "utlFDCAN_Task.h"
 
-#define LED_Status1_Pin GPIO_PIN_7
-#define LED_Status1_GPIO_Port GPIOC
-
-extern QueueHandle_t xQueueToUSB;
-extern QueueHandle_t xQueueToUART;
 extern void Error_Handler(void);
 
 osThreadId_t utlFDCAN_Task_Create(osThreadFunc_t function, void * argument, osThreadAttr_t * attr) {
@@ -40,24 +35,3 @@ void utlFDCAN_Task_Start(void * argument) {
   }
 }
 
-
-void HAL_FDCAN_RxFifo0Callback(FDCAN_HandleTypeDef *hfdcan, uint32_t RxFifo0ITs) {
-  static utlFDCAN_Data_t Data;
-  static FDCAN_RxHeaderTypeDef RxHeader;
-  static HAL_StatusTypeDef LastStatus_HAL = HAL_ERROR;
-  static BaseType_t xLastStatus_Os = pdFALSE;
-  static BaseType_t xHigherPriorityTaskWoken = pdFALSE;
-
-  LastStatus_HAL = HAL_FDCAN_GetRxMessage(hfdcan, FDCAN_RX_FIFO0, &RxHeader, (uint8_t*)&Data);
-  if (LastStatus_HAL != HAL_OK) {
-    Error_Handler();
-  }
-
-  xLastStatus_Os = xQueueSendToBackFromISR(xQueueToUART, &Data, &xHigherPriorityTaskWoken);
-  if (xLastStatus_Os != pdPASS) {
-    Error_Handler();
-  }
-
-  HAL_GPIO_TogglePin(LED_Status1_GPIO_Port, LED_Status1_Pin);
-
-}
